Cambiados tipos id a Long

+++ src/main/java/com/atm/buenas_practicas_java/entities/Amistad.java
@@ -21,7 +21,7 @@ public class Amistad {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int idAmistad;
+    private Long idAmistad;
     private boolean estado;
     private Date fecha;
