+++ src/main/java/com/atm/buenas_practicas_java/entities/UsuarioComunidad.java
@@ -15,7 +15,7 @@ import lombok.Setter;
 public class UsuarioComunidad {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idUsuarioComunidad;
+    private Long idUsuarioComunidad;
 
     @ManyToOne
     @JoinColumn(name = "id_usuario")
